requirements.txt
anyascii==0.3.0
appdirs==1.4.3
arrow==1.2.0
asgiref==3.4.1
asn1crypto==0.24.0
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
black==20.8b1
bleach==3.3.1
blinker==1.4
bottle==0.12.17
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
crispy-bootstrap5==0.4
cryptography==2.7
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.3
Django==3.2.8
django-appconf==1.0.3
django-autoslug==1.9.4
django-cogwheels==0.3
django-contrib-comments==2.1.0
django-cors-headers==3.6.0
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-filter==2.2.0
django-formtools==2.1
django-guardian==2.2.0
django-import-export==2.0.2
django-jsonfield==1.4.1
django-mdeditor==0.1.18
django-modelcluster==5.1
django-recaptcha==2.0.6
django-simpleui==2021.8.17
django-taggit==1.5.1
django-treebeard==4.5.1
django-utils-six==2.0
djangorestframework==3.12.2
draftjs-exporter==2.1.7
emoji==0.6.0
et-xmlfile==1.0.1
filelock==3.3.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-DebugToolbar==0.11.0
Flask-JWT-Extended==3.25.0
Flask-Script==2.0.6
flask-sqlacodegen==1.1.8
Flask-SQLAlchemy==2.4.4
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
httplib2==0.17.0
idna==2.9
importlib-metadata==3.4.0
inflect==5.0.2
iniconfig==1.1.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.1
jinja2-time==0.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
l18n==2020.6.1
lazy-object-proxy==1.4.1
Lin-CMS==0.3.0a9
Markdown==3.2.1
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
MouseInfo==0.1.3
mypy-extensions==0.4.3
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
packaging==20.8
parso==0.8.1
pathspec==0.8.1
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==2.4.0
pluggy==0.13.1
poyo==0.5.0
prompt-toolkit==3.0.17
psycopg2-binary==2.9.1
py==1.10.0
PyAutoGUI==0.9.52
pycparser==2.19
pydantic==1.7.3
PyGetWindow==0.0.9
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.6.0
PyMsgBox==1.0.9
PyMySQL==0.9.3
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.13.0
PyQt5-sip==12.7.1
pyqt5-tools==5.13.0.1.5
PyQtChart==5.14.0
PyQtWebEngine==5.14.0
PyRect==0.1.4
PyScreeze==0.1.26
pytest==6.2.1
pytest-ordering==0.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-slugify==5.0.2
python-webpack-boilerplate==0.0.3
PyTweening==1.0.3
pytz==2021.3
pywin32==227
PyYAML==5.3
pyzmq==22.0.3
QScintilla==2.11.4
regex==2020.11.13
requests==2.23.0
rjsmin==1.1.0
selenium==3.141.0
simplejson==3.17.2
six==1.12.0
soupsieve==2.2.1
spectree==0.3.16
SQLAlchemy==1.3.20
sqlparse==0.3.0
style==1.1.0
tablib==3.0.0
telepath==0.2
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
update==0.0.1
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.8.1
wagtail==2.14.1
wagtail-django-recaptcha==1.0
wagtail-markdown==0.7.0
wagtail-metadata==3.4.0
wagtailmenus==3.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
Willow==1.4
wrapt==1.11.2
WTForms==2.3.3
xlrd==1.2.0
XlsxWriter==1.4.5
xlwt==1.3.0
zipp==3.4.0
django-mdeditor
mysqlclient
pillow
alibabacloud_dysmsapi20170525==2.0.3